requirements.txt
ultralytics
Flask>=3.0.0
Pillow>=10.0.0
numpy>=1.23.0
streamlit
watchdog
